// src/types.ts
export type Slide = {
    key: string;
    variant: string;
    titleHtml: string;
    punchline: string;
    name?: string;
    subtitle?: string;
};

export type Translation = {
    features: string;
    signUp: string;
    instructions: string;
    prototypeNotes: string;
    note1: string;
    note2: string;
    note3: string;
    dashboardTitle: string;
    dashboardSubtitle: string;
    autopilotTitle: string;
    autopilotDesc: string;
    monthlyStats: string;
    monthlyStatsDesc: string;
    postsMade: string;
    totalSales: string;
    audienceGrowth: string;
    pendingOrders: string;
    order: string;
    price: string;
    pickup: string;
    upcomingEvents: string;
    liveWorkshop: string;
    onlineTutorial: string;
    trendingTopics: string;
    aiSuggestion: string;
    aiSuggestionText: string;
    socialEngagement: string;
    shares: string;
    comments: string;
    inquiries: string;
    off: string;
    on: string;
    autopilotEngaged: string;
    autopilotStandby: string;
    footerRights: string;
    featuresTitle: string;
    featuresSubtitle: string;
    creativeStudio: string;
    creativeStudioDesc: string;
    artworkGallery: string;
    artworkGalleryDesc: string;
    growthWallet: string;
    growthWalletDesc: string;
    hostWorkshop: string;
    hostWorkshopDesc: string;
    slides: Slide[];
};